+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Fixed
+
+- `send_contact` signature (`phone_number` and `first_name` `f64` => `String`) ([#56][pr56])
+
+[pr56]: https://github.com/teloxide/teloxide-core/pull/56
 
 ## [0.1.1] - 2020-02-17